WIP: Big ass action refactor still ongoing.

+++ Packages/com.unity.inputsystem/InputSystem/State/IInputStateChangeMonitor.cs
@@ -2,7 +2,7 @@ namespace UnityEngine.Experimental.Input.LowLevel
 {
     public interface IInputStateChangeMonitor
     {
-        void NotifyControlValueChanged(InputControl control, double time, int monitorIndex);
-        void NotifyTimerExpired(InputControl control, double time, int monitorIndex, int timerIndex);
+        void NotifyControlValueChanged(InputControl control, double time, long monitorIndex);
+        void NotifyTimerExpired(InputControl control, double time, long monitorIndex, int timerIndex);
     }
 }
